refactor: moving config into schemas module
